Fix model association. queue_type enum isn't working.
So, the model association bug was solved. The type column is used as a way to invoke the single
table inheritance. I changed the name from 'type'->'queue_type' to fix this. The enum for queue_type
isn't working for some reason though. If I have a Request instance of called request, and I enter
request.queue_type = "demo" I'll get 0. No matter what I put in it's 0. Tests still aren't working.
I think it's related to the queue_type enum not working.

 - fix model association
 - update migration
 - write test for request
 - update models